login/requirements.txt
flask
requests
Flask-SQLAlchemy==2.4.4
SQLAlchemy==1.3.20
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-Cors==3.0.9

Django==3.1.3
djangorestframework==3.12.2
django-mysql==3.9
django-cors-headers==3.5.0
psycopg2
